pyproject.toml
[tool.poetry]
name = "PDF_Video_Viewer"
version = "0.1.0"
description = ""
authors = ["sturgis <sturgis.steele@outlook.com>"]
readme = "README.md"
packages = []

[tool.poetry.dependencies]
python = "^3.12"
discord-py = "^2.5.2"
pdf2image = "^1.17.0"
google-api-python-client = "^2.166.0"
python-dotenv = "^1.1.0"
requests = "^2.32.3"
pillow = "^11.1.0"
pytz = "^2025.2"
pymupdf = "^1.25.4"
google-auth-oauthlib = "^1.2.1"
google-auth-httplib2 = "^0.2.0"
ffmpeg = "^1.4"
flask = "^3.1.0"
ffmpeg-python = "^0.2.0"
python-docx = "^1.1.2"
aiohttp = "^3.8.6"
beautifulsoup4 = "^4.12.2"
selenium = "^4.18.1"
webdriver-manager = "^4.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
